#2816: restructure, and reimplement database modules
