+++ brain/database/account.py
@@ -2,19 +2,18 @@
 
 '''
 
-This file retrieves user account values.
+This file queries the user account information.
 
 '''
 
 from flask import current_app
-from brain.database.db_query import SQL
+from brain.database.query import SQL
 
 
-class Retrieve_Account(object):
+class Account(object):
     '''
 
-    This class provides an interface to check if a username already exists,
-    and retrieves the corresponding password.
+    This class provides an interface to the users account.
 
     Note: this class explicitly inherits the 'new-style' class.
 
@@ -31,6 +30,36 @@ class Retrieve_Account(object):
         self.sql = SQL()
         self.db_ml = current_app.config.get('DB_ML')
 
+    def save_account(self, username, email, password):
+        '''
+
+        This method stores a user account, along with their corresponding
+        password into an 'EAV data model' database table.
+
+        @sql_statement, is a sql format string, and not a python string.
+            Therefore, '%s' is used for argument substitution.
+
+        '''
+
+        # insert / update dataset entity value
+        self.sql.connect(self.db_ml)
+
+        sql_statement = 'INSERT INTO tbl_user '\
+            '(username, email, password, datetime_joined) '\
+            'VALUES(%s, %s, %s, UTC_TIMESTAMP())'
+        args = (username, email, password)
+        response = self.sql.execute(sql_statement, 'insert', args)
+
+        # retrieve any error(s), disconnect from database
+        response_error = self.sql.get_errors()
+        self.sql.disconnect()
+
+        # return result
+        if response_error:
+            return {'status': False, 'error': response_error}
+        else:
+            return {'status': True, 'error': None, 'id': response['id']}
+
     def check_username(self, username):
         '''
 
@@ -39,16 +68,16 @@ class Retrieve_Account(object):
         '''
 
         # select dataset
-        self.sql.sql_connect(self.db_ml)
+        self.sql.connect(self.db_ml)
         sql_statement = 'SELECT * '\
             'FROM tbl_user '\
             'WHERE username=%s'
         args = (username)
-        response = self.sql.sql_command(sql_statement, 'select', args)
+        response = self.sql.execute(sql_statement, 'select', args)
 
         # retrieve any error(s), disconnect from database
         response_error = self.sql.get_errors()
-        self.sql.sql_disconnect()
+        self.sql.disconnect()
 
         # return result
         if response_error:
@@ -64,16 +93,16 @@ class Retrieve_Account(object):
         '''
 
         # select dataset
-        self.sql.sql_connect(self.db_ml)
+        self.sql.connect(self.db_ml)
         sql_statement = 'SELECT * '\
             'FROM tbl_user '\
             'WHERE email=%s'
         args = (email)
-        response = self.sql.sql_command(sql_statement, 'select', args)
+        response = self.sql.execute(sql_statement, 'select', args)
 
         # retrieve any error(s), disconnect from database
         response_error = self.sql.get_errors()
-        self.sql.sql_disconnect()
+        self.sql.disconnect()
 
         # return result
         if response_error:
@@ -89,16 +118,16 @@ class Retrieve_Account(object):
         '''
 
         # select dataset
-        self.sql.sql_connect(self.db_ml)
+        self.sql.connect(self.db_ml)
         sql_statement = 'SELECT password '\
             'FROM tbl_user '\
             'WHERE username=%s'
         args = (username)
-        response = self.sql.sql_command(sql_statement, 'select', args)
+        response = self.sql.execute(sql_statement, 'select', args)
 
         # retrieve any error(s), disconnect from database
         response_error = self.sql.get_errors()
-        self.sql.sql_disconnect()
+        self.sql.disconnect()
 
         # return result
         if response_error:
@@ -114,16 +143,16 @@ class Retrieve_Account(object):
         '''
 
         # select dataset
-        self.sql.sql_connect(self.db_ml)
+        self.sql.connect(self.db_ml)
         sql_statement = 'SELECT id_user '\
             'FROM tbl_user '\
             'WHERE username=%s'
         args = (username)
-        response = self.sql.sql_command(sql_statement, 'select', args)
+        response = self.sql.execute(sql_statement, 'select', args)
 
         # retrieve any error(s), disconnect from database
         response_error = self.sql.get_errors()
-        self.sql.sql_disconnect()
+        self.sql.disconnect()
 
         # return result
         if response_error:
