+++ brain/database/entity.py
@@ -2,28 +2,28 @@
 
 '''
 
-This file saves a dataset entity into corresponding 'EAV data model' database
-table(s), from the 'db_machine_learning' database.
+This file retrieves dataset entity related properties.
 
 '''
 
 from flask import current_app
-from brain.database.db_query import SQL
+from brain.database.query import SQL
 
 
-class Save_Entity(object):
+class Entity(object):
     '''
 
-    This class provides an interface to save dataset(s) later used to generate
-    corresponding model(s).
+    This class provides an interface to save, retrieve an SVM entity title,
+    from the 'tbl_dataset_entity' sql database table.
 
-    Note: this class is invoked within 'base_data.py', and 'data_append.py'
+    Note: this class is invoked within 'model_generate.py', 'base_data.py',
+          and 'data_append.py'
 
     Note: this class explicitly inherits the 'new-style' class.
 
     '''
 
-    def __init__(self, premodel_data, session_type):
+    def __init__(self, premodel_data=None, session_type=None):
         '''
 
         This constructor is responsible for defining class variables.
@@ -57,14 +57,14 @@ class Save_Entity(object):
         '''
 
         # insert / update dataset entity value
-        self.sql.sql_connect(self.db_ml)
+        self.sql.connect(self.db_ml)
 
         if self.session_type == 'data_append':
             sql_statement = 'UPDATE tbl_dataset_entity '\
                 'SET uid_modified=%s, datetime_modified=UTC_TIMESTAMP() '\
                 'WHERE id_entity=%s'
             args = (self.premodel_data['uid'], self.premodel_data['id_entity'])
-            response = self.sql.sql_command(sql_statement, 'update', args)
+            response = self.sql.execute(sql_statement, 'update', args)
 
         elif self.session_type == 'data_new':
             sql_statement = 'INSERT INTO tbl_dataset_entity '\
@@ -75,14 +75,46 @@ class Save_Entity(object):
                 self.premodel_data['model_type'],
                 self.premodel_data['uid']
             )
-            response = self.sql.sql_command(sql_statement, 'insert', args)
+            response = self.sql.execute(sql_statement, 'insert', args)
 
         # retrieve any error(s), disconnect from database
         response_error = self.sql.get_errors()
-        self.sql.sql_disconnect()
+        self.sql.disconnect()
 
         # return result
         if response_error:
             return {'status': False, 'error': response_error}
         else:
             return {'status': True, 'error': None, 'id': response['id']}
+
+    def get_title(self, id_entity):
+        '''
+
+        This method is responsible for retrieving an SVM entity title, from the
+        SQL database, using a fixed 'id_entity'.
+
+        @id_entity, this supplied argument corresponds to the 'id_entity'
+            column from the 'tbl_dataset_value' database table.
+
+        @sql_statement, is a sql format string, and not a python string.
+            Therefore, '%s' is used for argument substitution.
+
+        '''
+
+        # select dataset
+        self.sql.connect(self.db_ml)
+        sql_statement = 'SELECT title '\
+            'FROM tbl_dataset_entity '\
+            'WHERE id_entity=%s'
+        args = (id_entity)
+        response = self.sql.execute(sql_statement, 'select', args)
+
+        # retrieve any error(s), disconnect from database
+        response_error = self.sql.get_errors()
+        self.sql.disconnect()
+
+        # return result
+        if response_error:
+            return {'error': response_error, 'result': None}
+        else:
+            return {'error': None, 'result': response['result']}
