+++ test/live_server/authentication/pytest_account_registration.py
@@ -10,8 +10,7 @@ Note: the 'pytest' instances can further be reviewed:
 '''
 
 from brain.validator.validate_password import validate_password
-from brain.database.retrieve_account import Retrieve_Account
-from brain.database.save_account import Save_Account
+from brain.database.account import Account
 from brain.converter.crypto import hash_pass
 
 
@@ -28,7 +27,7 @@ def test_registration(client, live_server):
     username = 'jeff1evesque'
     email = 'jeff1evesque@yahoo.com'
     password = 'password123'
-    authenticate = Retrieve_Account()
+    authenticate = Account()
 
     # verify requirements: one letter, one number, and ten characters.
     if (validate_password(password)):
@@ -41,7 +40,7 @@ def test_registration(client, live_server):
 
                 # database query: save username, and password
                 hashed = hash_pass(str(password))
-                result = Save_Account().save_account(username, email, hashed)
+                result = Account().Account(username, email, hashed)
 
                 # notification: attempt to store account
                 assert result['status']
