+++ test/live_server/authentication/pytest_user_login.py
@@ -10,7 +10,7 @@ Note: the 'pytest' instances can further be reviewed:
 '''
 
 from flask import session
-from brain.database.retrieve_account import Retrieve_Account
+from brain.database.account import Account
 from brain.converter.crypto import verify_pass
 
 
@@ -30,7 +30,7 @@ def test_login(client, live_server):
     username = 'jeff1evesque'
     password = 'password123'
     url = '/login'
-    authenticate = Retrieve_Account()
+    authenticate = Account()
 
     # validate: username exists
     if authenticate.check_username(username)['result']:
